requirements.txt
Django==3.1.7
psycopg2-binary==2.8.5
dotenv-python==0.0.1
uwsgi
python-dotenv==0.14.0
djangorestframework==3.12.2
djangorestframework-simplejwt==4.6.0
django-cors-headers==3.5.0
django-environ==0.4.5
django_s3_storage==0.13.4
django-s3direct==1.1.8
Pillow==8.0.1
whitenoise